+++ src/onchain-collector/onchain-collector.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { OnchainCollectorService } from './onchain-collector.service';
+
+@Module({
+  providers: [OnchainCollectorService],
+  exports: [OnchainCollectorService],
+})
+export class OnchainCollectorModule {}